src/main/java/io/github/korzepadawid/springtaskplanning/service/impl/UserServiceImpl.java
package io.github.korzepadawid.springtaskplanning.service.impl;

import io.github.korzepadawid.springtaskplanning.dto.UserResponse;
import io.github.korzepadawid.springtaskplanning.exception.ResourceNotFoundException;
import io.github.korzepadawid.springtaskplanning.repository.UserRepository;
import io.github.korzepadawid.springtaskplanning.service.UserService;
import org.springframework.stereotype.Service;

@Service
public class UserServiceImpl implements UserService {

  private final UserRepository userRepository;

  public UserServiceImpl(UserRepository userRepository) {
    this.userRepository = userRepository;
  }

  @Override
  public UserResponse findUserByEmail(String email) {
    return userRepository
        .findByEmail(email)
        .map(UserResponse::new)
        .orElseThrow(() -> new ResourceNotFoundException("User doesn't exist"));
  }
}
